+++ packages/types/src/__tests__/index.test.ts
@@ -90,4 +90,6 @@ test('InferFaasAction should infer schema params from defineApi', () => {
 
   // @ts-expect-error key should be string
   assertType<InferredAction['Params']>({ key: 1 })
+
+  assertType<InferredAction['Data']>({ value: '' })
 })